+++ src/components/Button/Buttons.style.ts
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-const StyledButton = styled.button<{
+export const StyledButton = styled.button<{
   width: number;
   height: number;
   dark: boolean;
@@ -18,32 +18,3 @@ const StyledButton = styled.button<{
   border-radius: 10px;
   font-size: 1rem;
 `;
-
-interface ButtonProps {
-  width: number;
-  height: number;
-  dark: boolean;
-  label?: string;
-  handleClick?: () => void;
-  bold: boolean;
-}
-
-const Button = ({
-  width,
-  height,
-  label,
-  handleClick,
-  dark,
-  bold
-}: ButtonProps) => (
-  <StyledButton
-    width={width}
-    height={height}
-    onClick={handleClick}
-    dark={dark}
-    bold={bold}>
-    {label}
-  </StyledButton>
-);
-
-export default Button;
